requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
Pillow==8.1.0
pylint==2.6.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
